pyproject.toml
[tool.poetry]
name = "online-training"
version = "0.1.0"
description = ""
authors = ["Deadmau5"]
readme = "README.md"
packages = [{include = "online_training"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
psycopg2-binary = "^2.9.7"
pillow = "^10.0.1"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
ipython = "^8.15.0"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
